Use new dotenv-deployment conventions.
The new dotenv-deployment gem looks for .env-[environment] and calls
Dotenv.overload if it exists. This eliminates the need for us to
call overload explicitly in unicorn.rb.

Update the dotenv recipe to use the .env-[environment] filename
rather than the plain .env (the latter does not get the automatic
overload behavior).