Avoid invoking the same task hook multiple times
When the same task is registered as a before/after hook in multiple
places, that would a "doubling up" of that task definition, such that
each time it was called it would use `invoke` twice.

This was not causing any bugs, because Rake's behavior is to silently
skip invocations after a task has already been invoked once.

However Capistrano will print a warning starting with Capistrano 3.6.0,
so it is important to fix this mistake now.