Fix "tasks may only be invoked once" warning
When running `deploy:migrate_and_restart`, Sidekiq is first stopped,
then later the `mb:sidekiq:restart` task is used to gracefully start it
again. This results in the `mb:sidekiq:stop` task being invoked twice,
which is not permitted; newer versions of Capistrano show a warning.

Fix the problem and eliminate the warning by explicitly re-enabling the
`stop` task so that we can safely invoke it a second time.