Fix sidekiq + x-sendfile config
Sidekiq's assets (frustratingly) are located inside its gem, not in
public/assets like everything else. This requires they be served by the
Rails process. Furthermore, Rails makes use of x-sendfile to deliver
these files quickly. But since the Sidekiq files are in a gem, they live
in the Capistrano shared location, not in the app. That means we have to
relax the Nginx x-sendfile configuration to allow serving *all* files
within :deploy_to, not just the releases.